Timeline and form changes. See description.
Made timeline page appear when user is loged in.

Added logout button and the current user to the display. I didn't have a chance to style them.

I also really like that Rose created a new task form in another file, so I decided to do that with the rest of our task forms and put them in a folder.