Progress on canvas drawing
We now have the date from the database reading out and correctly comparing with the Start Day selected. We need to yet 1) fill out a draw function that creates an appropriate task line. 2) make if statements that figure out circumstances where the event should be on screen, but is not specifically starting on the Start Date selected.