Figured out how to correctly set the dateVal attribute for both the lTermTimeSlot spans in the lTermTimeIncColHeader and the sTermTimeSlot spans in the sTermTimeIncColHeader using moment.js and timeline lines displayed mostly accurately for all events in database. There is still a bug in displaying events that span multiple years.
There is still a bug for events that span multiple years. For example, the event called "Test task spanning several years" is supposed to last from 7/1/2017 - 7/1/2019. The ending dot is not displayed in short term view. The starting dot does not show up in the short term view; the line representing the timeline line comes out of the canvas. For the long term view, when a date in June 2017 is input and long term view is clicked, the event appears to start in June 2017, as there is a starting dot there; however, the event actually starts in July 2017.